Syntax can require object flags
Users can now enter syntax that requires the objects entered to meet some critera before executing any game actions. For example, if the player enters 'look at the dog', and the syntax specifies that the 'look at [OBJECT]' command requires the object to be in the room or in the inventory, the 'look at' function will not be run, and the a suitable message will be displayed, such as 'You can't see any dog here'.