Prototype completed
The prototype for azkii is now complete. An example game has been added, along with all its data tables. Many bugs have also been fixed. There are still some things to be improved on, but these will most likely be implemented in the final program, which will be in unity.